fixed quadrapong
