+++ pettingzoo/atari/pong_quadrapong_v1.py
@@ -1,0 +1,11 @@
+from .base_atari_env import BaseAtariEnv, base_env_wrapper_fn, parallel_wrapper_fn
+
+
+def raw_env(**kwargs):
+    mode = 33
+    num_players = 4
+    return BaseAtariEnv(game="pong", num_players=num_players, mode_num=mode, **kwargs)
+
+
+env = base_env_wrapper_fn(raw_env)
+parallel_env = parallel_wrapper_fn(env)
